third_party/blink/web_tests/fast/hidpi/image-srcset-dpr-zoom.html

<!DOCTYPE html>
<script>
    if (!window.testRunner) {
        document.write("Automatic zoom-in does not work in manual mode. Try zooming in manually and refresh.");
    } else {
        while (!matchMedia("screen and (min-resolution: 2dppx)").matches)
            testRunner.zoomPageIn();
    }
</script>
<script src="../../resources/js-test.js"></script>
<script>
    function updateSrc() {
        var img = document.getElementById("foo");
        img.srcset = "resources/blue-100-px-square.png 1x, resources/green-400-px-square.png 2x";
    }
    addEventListener("DOMContentLoaded", function() {
        updateSrc();
    }, false);
    addEventListener("load", function() {
        shouldBeTrue('Math.abs(document.getElementById("foo").clientWidth - 200) < 0.1');
    }, false);
</script>
<div>This test passes if the image below is a 200px wide green square when the deviceScaleFactor is 1, since the user has zoomed in.</div>
<img id="foo">
